+++ push_notifications/fields.py
@@ -1,29 +1,45 @@
 import re
 import struct
 from django import forms
+from django.core.validators import MaxValueValidator
+from django.core.validators import MinValueValidator
 from django.core.validators import RegexValidator
 from django.db import models, connection
+from django.utils import six
 from django.utils.translation import ugettext_lazy as _
 
-try:
-	from django.utils import six
-except ImportError:
-	import six
+UNSIGNED_64BIT_INT_MIN_VALUE = 0
+UNSIGNED_64BIT_INT_MAX_VALUE = 2 ** 64 - 1
+
+__all__ = ["HexadecimalField", "HexIntegerField"]
 
 
-__all__ = ["HexadecimalField", "HexIntegerField", "UUIDField"]
-
-# Django <1.8 compatibility: UUIDField
-if hasattr(models, "UUIDField"):
-	UUIDField = models.UUIDField
-else:
-	from uuidfield import UUIDField
-
-hex_re = re.compile(r"^0x[0-9a-fA-F]+$")
-postgres_engines = [
+hex_re = re.compile(r"^(([0-9A-f])|(0x[0-9A-f]))+$")
+signed_integer_engines = [
 	"django.db.backends.postgresql_psycopg2",
 	"django.contrib.gis.db.backends.postgis",
+	"django.db.backends.sqlite3"
 ]
+
+
+def _using_signed_storage():
+	return connection.settings_dict["ENGINE"] in signed_integer_engines
+
+
+def _signed_to_unsigned_integer(value):
+	return struct.unpack("Q", struct.pack("q", value))[0]
+
+
+def _unsigned_to_signed_integer(value):
+	return struct.unpack("q", struct.pack("Q", value))[0]
+
+
+def _hex_string_to_unsigned_integer(value):
+	return int(value, 16)
+
+
+def _unsigned_integer_to_hex_string(value):
+	return hex(value).rstrip("L")
 
 
 class HexadecimalField(forms.CharField):
@@ -34,8 +50,15 @@ class HexadecimalField(forms.CharField):
 		self.default_validators = [RegexValidator(hex_re, _("Enter a valid hexadecimal number"), "invalid")]
 		super(HexadecimalField, self).__init__(*args, **kwargs)
 
+	def prepare_value(self, value):
+		# converts bigint from db to hex before it is displayed in admin
+		if value and not isinstance(value, six.string_types) \
+			and connection.vendor in ("mysql", "sqlite"):
+			value = _unsigned_integer_to_hex_string(value)
+		return super(forms.CharField, self).prepare_value(value)
 
-class HexIntegerField(six.with_metaclass(models.SubfieldBase, models.BigIntegerField)):
+
+class HexIntegerField(models.BigIntegerField):
 	"""
 	This field stores a hexadecimal *string* of up to 64 bits as an unsigned integer
 	on *all* backends including postgres.
@@ -47,34 +70,46 @@ class HexIntegerField(six.with_metaclass(models.SubfieldBase, models.BigIntegerF
 	On sqlite and mysql, native unsigned bigint types are used. In all cases, the
 	value we deal with in python is always in hex.
 	"""
+
+	validators = [
+		MinValueValidator(UNSIGNED_64BIT_INT_MIN_VALUE),
+		MaxValueValidator(UNSIGNED_64BIT_INT_MAX_VALUE)
+	]
+
 	def db_type(self, connection):
 		engine = connection.settings_dict["ENGINE"]
-		if engine == "django.db.backends.mysql":
+		if "mysql" in engine:
 			return "bigint unsigned"
-		elif engine == "django.db.backends.sqlite":
+		elif "sqlite" in engine:
 			return "UNSIGNED BIG INT"
 		else:
 			return super(HexIntegerField, self).db_type(connection=connection)
 
 	def get_prep_value(self, value):
+		""" Return the integer value to be stored from the hex string """
 		if value is None or value == "":
 			return None
 		if isinstance(value, six.string_types):
-			value = long(value, 16)
-		# on postgres only, interpret as signed
-		if connection.settings_dict["ENGINE"] in postgres_engines:
-			value = struct.unpack("q", struct.pack("Q", value))[0]
+			value = _hex_string_to_unsigned_integer(value)
+		if _using_signed_storage():
+			value = _unsigned_to_signed_integer(value)
+		return value
+
+	def from_db_value(self, value, expression, connection, context):
+		""" Return an unsigned int representation from all db backends """
+		if value is None:
+			return value
+		if _using_signed_storage():
+			value = _signed_to_unsigned_integer(value)
 		return value
 
 	def to_python(self, value):
+		""" Return a str representation of the hexadecimal """
 		if isinstance(value, six.string_types):
 			return value
 		if value is None:
-			return ""
-		# on postgres only, re-interpret from signed to unsigned
-		if connection.settings_dict["ENGINE"] in postgres_engines:
-			value = hex(struct.unpack("Q", struct.pack("q", value))[0])
-		return value
+			return value
+		return _unsigned_integer_to_hex_string(value)
 
 	def formfield(self, **kwargs):
 		defaults = {"form_class": HexadecimalField}
@@ -82,8 +117,7 @@ class HexIntegerField(six.with_metaclass(models.SubfieldBase, models.BigIntegerF
 		# yes, that super call is right
 		return super(models.IntegerField, self).formfield(**defaults)
 
-try:
-	from south.modelsinspector import add_introspection_rules
-	add_introspection_rules([], ["^push_notifications\.fields\.HexIntegerField"])
-except ImportError:
-	pass
+	def run_validators(self, value):
+		# make sure validation is performed on integer value not string value
+		value = _hex_string_to_unsigned_integer(value)
+		return super(models.BigIntegerField, self).run_validators(value)