+++ push_notifications/gcm.py
@@ -48,18 +48,66 @@ def _gcm_send(data, content_type):
 	}
 
 	request = Request(SETTINGS["GCM_POST_URL"], data, headers)
+
 	try:
-		return urlopen(request).read()
+		return urlopen(request).read().decode("utf-8")
 	except HTTPError as e:
 		logger.error(
 			u"Gcm -> Error {} with data {} headers {}".format(
 				unicode(e), unicode(data), unicode(headers)
 			)
-		)
+)
 		raise e
 
 
-def _gcm_send_json(registration_ids, data, collapse_key=None, delay_while_idle=False, time_to_live=0):
+def _gcm_send_plain(registration_id, data, **kwargs):
+	"""
+	Sends a GCM notification to a single registration_id.
+	This will send the notification as form data.
+	If sending multiple notifications, it is more efficient to use
+	gcm_send_bulk_message() with a list of registration_ids
+	"""
+
+	values = {"registration_id": registration_id}
+
+	for k, v in data.items():
+		values["data.%s" % (k)] = v.encode("utf-8")
+
+	for k, v in kwargs.items():
+		if v:
+			if isinstance(v, bool):
+				# Encode bools into ints
+				v = 1
+			values[k] = v
+
+	data = urlencode(sorted(values.items())).encode("utf-8")  # sorted items for tests
+
+	result = _gcm_send(data, "application/x-www-form-urlencoded;charset=UTF-8")
+
+	# Information about handling response from Google docs (https://developers.google.com/cloud-messaging/http):
+	# If first line starts with id, check second line:
+	# If second line starts with registration_id, gets its value and replace the registration tokens in your
+	# server database. Otherwise, get the value of Error
+
+	if result.startswith("id"):
+		lines = result.split("\n")
+		if len(lines) > 1 and lines[1].startswith("registration_id"):
+			new_id = lines[1].split("=")[-1]
+			_gcm_handle_canonical_id(new_id, registration_id)
+
+	elif result.startswith("Error="):
+		if result in ("Error=NotRegistered", "Error=InvalidRegistration"):
+			# Deactivate the problematic device
+			device = GCMDevice.objects.filter(registration_id=values["registration_id"])
+			device.update(active=0)
+			return result
+
+		raise GCMError(result)
+
+	return result
+
+
+def _gcm_send_json(registration_ids, data, **kwargs):
 	"""
 	Sends a GCM notification to one or more registration_ids. The registration_ids
 	needs to be a list.
@@ -67,50 +115,86 @@ def _gcm_send_json(registration_ids, data, collapse_key=None, delay_while_idle=F
 	"""
 
 	if not registration_ids:
-                return
+		return
 
 	values = {"registration_ids": registration_ids}
 
 	if data is not None:
 		values["data"] = data
 
-	if collapse_key:
-		values["collapse_key"] = collapse_key
-
-	if delay_while_idle:
-		values["delay_while_idle"] = delay_while_idle
-
-	if time_to_live:
-		values["time_to_live"] = time_to_live
+	for k, v in kwargs.items():
+		if v:
+			values[k] = v
 
 	data = json.dumps(values, separators=(",", ":"), sort_keys=True).encode("utf-8")  # keys sorted for tests
 
-	result = json.loads(_gcm_send(data, "application/json"))
-	return _cm_handle_response(registration_ids, result)
+	response = json.loads(_gcm_send(data, "application/json"))
+	if response["failure"] or response["canonical_ids"]:
+		ids_to_remove, old_new_ids = [], []
+		throw_error = False
+		for index, result in enumerate(response["results"]):
+			error = result.get("error")
+			if error:
+				# Information from Google docs (https://developers.google.com/cloud-messaging/http)
+				# If error is NotRegistered or InvalidRegistration, then we will deactivate devices because this
+				# registration ID is no more valid and can't be used to send messages, otherwise raise error
+				if error in ("NotRegistered", "InvalidRegistration"):
+					ids_to_remove.append(registration_ids[index])
+				else:
+					throw_error = True
+
+			# If registration_id is set, replace the original ID with the new value (canonical ID) in your
+			# server database. Note that the original ID is not part of the result, so you need to obtain it
+			# from the list of registration_ids passed in the request (using the same index).
+			new_id = result.get("registration_id")
+			if new_id:
+				old_new_ids.append((registration_ids[index], new_id))
+
+		if ids_to_remove:
+			removed = GCMDevice.objects.filter(registration_id__in=ids_to_remove)
+			removed.update(active=0)
+
+		for old_id, new_id in old_new_ids:
+			_gcm_handle_canonical_id(new_id, old_id)
+
+		if throw_error:
+			raise GCMError(response)
+	return response
 
 
-def gcm_send_message(registration_id, data, collapse_key=None, delay_while_idle=False, time_to_live=0):
+def _gcm_handle_canonical_id(canonical_id, current_id):
+	"""
+	Handle situation when GCM server response contains canonical ID
+	"""
+	if GCMDevice.objects.filter(registration_id=canonical_id, active=True).exists():
+		GCMDevice.objects.filter(registration_id=current_id).update(active=False)
+	else:
+		GCMDevice.objects.filter(registration_id=current_id).update(registration_id=canonical_id)
+
+
+def gcm_send_message(registration_id, data, **kwargs):
 	"""
 	Sends a GCM notification to a single registration_id.
 
-	This will send the notification as form data if possible, otherwise it will
-	fall back to json data.
-
 	If sending multiple notifications, it is more efficient to use
 	gcm_send_bulk_message() with a list of registration_ids
+
+	A reference of extra keyword arguments sent to the server is available here:
+	https://developers.google.com/cloud-messaging/server-ref#downstream
 	"""
 
-	args = data, collapse_key, delay_while_idle, time_to_live
-	_gcm_send_json([registration_id], *args)
+	return _gcm_send_plain(registration_id, data, **kwargs)
 
-def gcm_send_bulk_message(registration_ids, data, collapse_key=None, delay_while_idle=False, time_to_live=0):
+
+def gcm_send_bulk_message(registration_ids, data, **kwargs):
 	"""
 	Sends a GCM notification to one or more registration_ids. The registration_ids
 	needs to be a list.
 	This will send the notification as json data.
-	"""
 
-	args = data, collapse_key, delay_while_idle, time_to_live
+	A reference of extra keyword arguments sent to the server is available here:
+	https://developers.google.com/cloud-messaging/server-ref#downstream
+	"""
 
 	# GCM only allows up to 1000 reg ids per bulk message
 	# https://developer.android.com/google/gcm/gcm.html#request
@@ -118,55 +202,7 @@ def gcm_send_bulk_message(registration_ids, data, collapse_key=None, delay_while
 	if len(registration_ids) > max_recipients:
 		ret = []
 		for chunk in _chunks(registration_ids, max_recipients):
-			ret.append(_gcm_send_json(chunk, *args))
+			ret.append(_gcm_send_json(chunk, data, **kwargs))
 		return ret
 
-	return _gcm_send_json(registration_ids, *args)
-
-def _cm_handle_response(registration_ids, response_data):
-	response = response_data
-	if response.get("failure") or response.get("canonical_ids"):
-		ids_to_remove, old_new_ids = [], []
-		throw_error = False
-		for index, result in enumerate(response["results"]):
-			error = result.get("error")
-			if error:
-				# https://firebase.google.com/docs/cloud-messaging/http-server-ref#error-codes
-				# If error is NotRegistered or InvalidRegistration, then we will deactivate devices
-				# because this registration ID is no more valid and can't be used to send messages,
-				# otherwise raise error
-				if error in ("NotRegistered", "InvalidRegistration"):
-					ids_to_remove.append(registration_ids[index])
-				else:
-					throw_error = True
-
-			# If registration_id is set, replace the original ID with the new value (canonical ID)
-			# in your server database. Note that the original ID is not part of the result, you need
-			# to obtain it from the list of registration_ids in the request (using the same index).
-			new_id = result.get("registration_id")
-			if new_id:
-				old_new_ids.append((registration_ids[index], new_id))
-
-		if ids_to_remove:
-			removed = GCMDevice.objects.filter(
-				registration_id__in=ids_to_remove
-			)
-			removed.update(active=0)
-
-		for old_id, new_id in old_new_ids:
-			_cm_handle_canonical_id(new_id, old_id)
-
-		if throw_error:
-			raise GCMError(response)
-	return response
-
-def _cm_handle_canonical_id(canonical_id, current_id):
-	"""
-	Handle situation when FCM server response contains canonical ID
-	"""
-	devices = GCMDevice.objects
-	if devices.filter(registration_id=canonical_id, active=True).exists():
-		devices.filter(registration_id=current_id).update(active=False)
-	else:
-		devices.filter(registration_id=current_id).update(registration_id=canonical_id)
-
+	return _gcm_send_json(registration_ids, data, **kwargs)