+++ CHANGELOG.rst
@@ -1,3 +1,40 @@
+v1.4.1 (2016-01-11)
+===================
+* APNS: Increased max device token size to 100 bytes (WWDC 2015, iOS 9)
+* BUGFIX: Fix an index error in the admin
+
+v1.4.0 (2015-12-13)
+===================
+* BACKWARDS-INCOMPATIBLE: Drop support for Python<3.4
+* DJANGO: Support Django 1.9
+* GCM: Handle canonical IDs
+* GCM: Allow full range of GCMDevice.device_id values
+* GCM: Do not allow duplicate registration_ids
+* DRF: Work around empty boolean defaults issue (django-rest-framework#1101)
+* BUGFIX: Do not throw GCMError in bulk messages from the admin
+* BUGFIX: Avoid generating an extra migration on Python 3
+* BUGFIX: Only send in bulk to active devices
+* BUGFIX: Display models correctly in the admin on both Python 2 and 3
+
+
+v1.3.1 (2015-06-30)
+===================
+This is an errata release.
+
+v1.3.0 (2015-06-30)
+===================
+* BACKWARDS-INCOMPATIBLE: Drop support for Python<2.7
+* BACKWARDS-INCOMPATIBLE: Drop support for Django<1.8
+* NEW FEATURE: Added a Django Rest Framework API. Requires DRF>=3.0.
+* APNS: Add support for setting the ca_certs file with new APNS_CA_CERTIFICATES setting
+* GCM: Deactivate GCMDevices when their notifications cause NotRegistered or InvalidRegistration
+* GCM: Indiscriminately handle all keyword arguments in gcm_send_message and gcm_send_bulk_message
+* GCM: Never fall back to json in gcm_send_message
+* BUGFIX: Fixed migration issues from 1.2.0 upgrade.
+* BUGFIX: Better detection of SQLite/GIS MySQL in various checks
+* BUGFIX: Assorted Python 3 bugfixes
+* BUGFIX: Fix display of device_id in admin
+
 v1.2.1 (2015-04-11)
 ===================
 * APNS, GCM: Add a db_index to the device_id field